Refactor app structure: separate pages, widgets, models, and restore CategorySection functionality
