+++ lib/models/lesson_item.dart
@@ -1,0 +1,13 @@
+class LessonItem {
+  final String title;
+  final String description;
+  final String difficulty;
+  final String duration;
+
+  const LessonItem({
+    required this.title,
+    required this.description,
+    required this.difficulty,
+    required this.duration,
+  });
+}
